+++ front/change.form.php
@@ -196,6 +196,13 @@ if (isset($_POST["add"])) {
     );
     Html::redirect(Change::getFormURLWithID($id));
 } else {
+    // Add a change from item : format data
+    if (
+        isset($_REQUEST['_add_fromitem'], $_REQUEST['itemtype'], $_REQUEST['items_id'])
+    ) {
+        $_REQUEST['items_id'] = [$_REQUEST['itemtype'] => [$_REQUEST['items_id']]];
+    }
+
     if (isset($_GET['showglobalkanban']) && $_GET['showglobalkanban']) {
         Html::header(sprintf(__('%s Kanban'), Change::getTypeName(1)), $_SERVER['PHP_SELF'], "helpdesk", "change");
         $change::showKanban(0);