+++ front/item_problem.form.php
@@ -33,28 +33,8 @@
  * ---------------------------------------------------------------------
  */
 
-use Glpi\Event;
-
 include('../inc/includes.php');
 
-Session::checkLoginUser();
-
-$item = new Item_Problem();
-
-if (isset($_POST["add"])) {
-    $item->check(-1, CREATE, $_POST);
-
-    if ($item->add($_POST)) {
-        Event::log(
-            $_POST["problems_id"],
-            "problem",
-            4,
-            "tracking",
-            //TRANS: %s is the user login
-            sprintf(__('%s adds a link with an item'), $_SESSION["glpiname"])
-        );
-    }
-    Html::back();
-}
-
-Html::displayErrorAndDie("lost");
+$obj      = new Problem();
+$item_obj = new Item_Problem();
+include(GLPI_ROOT . '/front/commonitilobject_item.form.php');