+++ src/Change_Item.php
@@ -33,8 +33,6 @@
  * ---------------------------------------------------------------------
  */
 
-use Glpi\Asset\Asset;
-
 /**
  * Change_Item Class
  *
@@ -82,145 +80,6 @@ class Change_Item extends CommonItilObject_Item
             return false;
         }
         return parent::prepareInputForAdd($input);
-    }
-
-
-    /**
-     * Print the HTML array for Items linked to a change
-     *
-     * @param $change Change object
-     *
-     * @return boolean|void
-     **/
-    public static function showForChange(Change $change)
-    {
-
-        $instID = $change->fields['id'];
-
-        if (!$change->can($instID, READ)) {
-            return false;
-        }
-        $canedit = $change->canEdit($instID);
-        $rand    = mt_rand();
-
-        $types_iterator = self::getDistinctTypes($instID);
-        $number = count($types_iterator);
-
-        if ($canedit) {
-            echo "<div class='firstbloc'>";
-            echo "<form name='changeitem_form$rand' id='changeitem_form$rand' method='post'
-                action='" . Toolbox::getItemTypeFormURL(__CLASS__) . "'>";
-
-            echo "<table class='tab_cadre_fixe'>";
-            echo "<tr class='tab_bg_2'><th colspan='2'>" . __('Add an item') . "</th></tr>";
-
-            echo "<tr class='tab_bg_1'><td>";
-            $types = [];
-            foreach (array_keys($change->getAllTypesForHelpdesk()) as $key) {
-                $types[] = $key;
-            }
-            Dropdown::showSelectItemFromItemtypes(['itemtypes'
-                                                      => $types,
-                'entity_restrict'
-                                                      => ($change->fields['is_recursive']
-                                                          ? getSonsOf(
-                                                              'glpi_entities',
-                                                              $change->fields['entities_id']
-                                                          )
-                                                          : $change->fields['entities_id'])
-            ]);
-            echo "</td><td class='center' width='30%'>";
-            echo "<input type='submit' name='add' value=\"" . _sx('button', 'Add') . "\" class='btn btn-primary'>";
-            echo "<input type='hidden' name='changes_id' value='$instID'>";
-            echo "</td></tr>";
-            echo "</table>";
-            Html::closeForm();
-            echo "</div>";
-        }
-
-        echo "<div class='spaced'>";
-        if ($canedit && $number) {
-            Html::openMassiveActionsForm('mass' . __CLASS__ . $rand);
-            $massiveactionparams = ['container' => 'mass' . __CLASS__ . $rand];
-            Html::showMassiveActions($massiveactionparams);
-        }
-
-        echo "<table class='tab_cadre_fixehov'>";
-        $header_begin  = "<tr>";
-        $header_top    = '';
-        $header_bottom = '';
-        $header_end    = '';
-        if ($canedit && $number) {
-            $header_top    .= "<th width='10'>" . Html::getCheckAllAsCheckbox('mass' . __CLASS__ . $rand);
-            $header_top    .= "</th>";
-            $header_bottom .= "<th width='10'>" . Html::getCheckAllAsCheckbox('mass' . __CLASS__ . $rand);
-            $header_bottom .= "</th>";
-        }
-        $header_end .= "<th>" . _n('Type', 'Types', 1) . "</th>";
-        $header_end .= "<th>" . Entity::getTypeName(1) . "</th>";
-        $header_end .= "<th>" . __('Name') . "</th>";
-        $header_end .= "<th>" . __('Serial number') . "</th>";
-        $header_end .= "<th>" . __('Inventory number') . "</th></tr>";
-        echo $header_begin . $header_top . $header_end;
-
-        foreach ($types_iterator as $type_row) {
-            $itemtype = $type_row['itemtype'];
-            if (!($item = getItemForItemtype($itemtype))) {
-                continue;
-            }
-            if ($item->canView()) {
-                $iterator = self::getTypeItems($instID, $itemtype);
-                $nb = count($iterator);
-
-                $prem = true;
-                foreach ($iterator as $data) {
-                    $link     = $itemtype::getFormURLWithID($data['id']);
-                    $linkname = $data["name"] ?? '';
-                    if (
-                        $_SESSION["glpiis_ids_visible"]
-                        || empty($data["name"])
-                    ) {
-                        $linkname = sprintf(__('%1$s (%2$s)'), $linkname, $data["id"]);
-                    }
-                    $name = "<a href=\"" . $link . "\">" . $linkname . "</a>";
-
-                    echo "<tr class='tab_bg_1'>";
-                    if ($canedit) {
-                        echo "<td width='10'>";
-                        Html::showMassiveActionCheckBox(__CLASS__, $data["linkid"]);
-                        echo "</td>";
-                    }
-                    if ($prem) {
-                        $itemname = $item->getTypeName($nb);
-                        echo "<td class='center top' rowspan='$nb'>" .
-                         ($nb > 1 ? sprintf(__('%1$s: %2$s'), $itemname, $nb) : $itemname) . "</td>";
-                        $prem = false;
-                    }
-                    echo "<td class='center'>";
-                    echo Dropdown::getDropdownName("glpi_entities", $data['entity']) . "</td>";
-                    echo "<td class='center" .
-                      (isset($data['is_deleted']) && $data['is_deleted'] ? " tab_bg_2_2'" : "'");
-                    echo ">" . $name . "</td>";
-                    echo "<td class='center'>" .
-                      (isset($data["serial"]) ? "" . $data["serial"] . "" : "-") . "</td>";
-                    echo "<td class='center'>" .
-                      (isset($data["otherserial"]) ? "" . $data["otherserial"] . "" : "-") . "</td>";
-                    echo "</tr>";
-                }
-            }
-        }
-
-        if ($number) {
-            echo $header_begin . $header_bottom . $header_end;
-        }
-
-        echo "</table>";
-        if ($canedit && $number) {
-            $massiveactionparams['ontop'] = false;
-            Html::showMassiveActions($massiveactionparams);
-            Html::closeForm();
-        }
-        echo "</div>";
     }
 
 
@@ -294,7 +153,7 @@ class Change_Item extends CommonItilObject_Item
 
         switch ($item->getType()) {
             case 'Change':
-                self::showForChange($item);
+                self::showForObject($item);
                 break;
 
             default: